Extract _redact.py + _file_utils.py from _legacy.py
